voice-search-darwin-arm64-docker/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==24.2.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datasets==2.14.4
dill==0.3.7
filelock==3.15.4
frozenlist==1.4.0
fsspec==2024.6.1
h2==4.1.0
hpack==4.0.0
huggingface_hub==0.24.5
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.3
nltk==3.8.2
numexpr==2.8.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.2
pyarrow==14.0.2
pycparser==2.22
pydub==0.25.1
PySocks==1.7.1
python-dateutil==2.9.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.1.99
setuptools==72.1.0
six==1.16.0
soundfile==0.12.1
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.18.1a0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
xxhash==2.0.2
yarl==1.9.3
zipp==3.20.0
zstandard==0.22.0
